requirements.txt
fastapi
uvicorn
pillow
numpy
torch
torchvision
torchaudio
python-multipart
tensorflow
scikit-learn
pydantic
joblib
pandas